cleaned up test files, condensed
