+++ api/tests/test_create_order.py
@@ -1,5 +1,6 @@
 from fastapi.testclient import TestClient
 import pytest
+from fastapi import HTTPException
 from unittest.mock import patch
 from main import app
 from queries.orders import OrderQueries
@@ -118,6 +119,34 @@ class TestUser:
 
             assert response.status_code == 422
             assert "detail" in response.json()
+
+            # Clean up dependency overrides
+            app.dependency_overrides = {}
+
+    def test_create_order_invalid_token(self, dummy_order, dummy_user):
+        with patch.object(
+            authenticator, "get_account_data_for_cookie", return_value=(dummy_user.username, dummy_user)), \
+            patch.object(OrderQueries, "create", new=CreateOrderQueries().create_mock_order):
+
+            def override_get_current_account_data():
+                raise HTTPException(status_code=401, detail="Invalid token")
+
+            app.dependency_overrides[authenticator.get_current_account_data] = override_get_current_account_data
+
+            token = "invalid_token"
+            headers = {"Authorization": f"Bearer {token}"}
+
+            json = {
+                "orders": [
+                    dummy_order.dict()
+                ]
+            }
+
+            response = client.post("/orders", json=json, headers=headers)
+            print(response.json())
+
+            assert response.status_code == 401
+            assert response.json() == {"detail": "Invalid token"}
 
             # Clean up dependency overrides
             app.dependency_overrides = {}
